% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_doc.R
\docType{data}
\name{alienista}
\alias{alienista}
\title{O Alienista - Machado de Assis}
\format{Um dataframe com 284 linhas e 5 variaveis:
\describe{
  \item{\code{book_name}}{character. Nome do livro.}
  \item{\code{chapter_name}}{character. Capitulo.}
  \item{\code{url}}{character. Link para capitulo no Wikisource.}
  \item{\code{paragraph_number}}{integer. Numero do paragrafo.}
  \item{\code{text}}{character. Texto do paragrafo.}
}}
\source{
\url{https://pt.wikisource.org/wiki/O_Alienista}
}
\usage{
alienista
}
\description{
Um dataset que contem o texto integral do livro O Alienista, de Machado de
Assis, publicado em 1882
}
\keyword{datasets}
