% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_doc.R
\docType{data}
\name{ateneu}
\alias{ateneu}
\title{O Ateneu - Raul Pompeia}
\format{Um dataframe com 1085 linhas e 5 variaveis:
\describe{
  \item{\code{book_name}}{character. Nome do livro.}
  \item{\code{chapter_name}}{character. Capitulo.}
  \item{\code{url}}{character. Link para capitulo no Wikisource.}
  \item{\code{paragraph_number}}{integer. Numero do paragrafo.}
  \item{\code{text}}{character. Texto do paragrafo.}
}}
\usage{
ateneu
}
\description{
Um dataset que contem o texto integral do livro O Ateneu, de Raul Pompeia,
publicado em 1888.
}
\keyword{datasets}
