% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_doc.R
\docType{data}
\name{cortico}
\alias{cortico}
\title{O Cortiço - Aluísio Azevedo}
\format{A data frame with 2232 rows and 5 variables:
\describe{
  \item{\code{book_name}}{character. Nome do livro.}
  \item{\code{chapter_name}}{character. Capitulo.}
  \item{\code{url}}{character. Link para capitulo no Wikisource.}
  \item{\code{paragraph_number}}{integer. Numero do paragrafo.}
  \item{\code{text}}{character. Texto do paragrafo.}
}}
\usage{
cortico
}
\description{
Um dataset que contem o texto integral do livro O Cortiço,
de Aluísio Azevedo,
publicado em 1890.
}
\keyword{datasets}
