% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_doc.R
\docType{data}
\name{dom_casmurro}
\alias{dom_casmurro}
\title{Dom Casmurro - Machado de Assis}
\format{A data frame with 612 rows and 5 variables:
\describe{
  \item{\code{book_name}}{character. Nome do livro.}
  \item{\code{chapter_name}}{character. Capitulo.}
  \item{\code{url}}{character. Link para capitulo no Wikisource.}
  \item{\code{paragraph_number}}{integer. Numero do paragrafo.}
  \item{\code{text}}{character. Texto do paragrafo.}
}}
\source{
\url{https://pt.wikisource.org/wiki/Dom_Casmurro}
}
\usage{
dom_casmurro
}
\description{
Título: Dom Casmurro
Autor: Machado de Assis
Ano: 1899
}
\keyword{datasets}
