% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_doc.R
\docType{data}
\name{escrava_isaura}
\alias{escrava_isaura}
\title{A Escrava Isaura - Bernardo Guimaraes}
\format{Um dataframe com 1293 linhas e 5 variaveis:
\describe{
  \item{\code{book_name}}{character. Nome do livro.}
  \item{\code{chapter_name}}{character. Capitulo.}
  \item{\code{url}}{character. Link para capitulo no Wikisource.}
  \item{\code{paragraph_number}}{integer. Numero do paragrafo.}
  \item{\code{text}}{character. Texto do paragrafo.}
}}
\source{
\url{https://pt.wikisource.org/wiki/A_escrava_Isaura}
}
\usage{
escrava_isaura
}
\description{
Um dataset que contem o texto integral do livro A Escrava Isaura, de Bernardo Guimaraes,
publicado em 1875.
}
\keyword{datasets}
