% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_all_books.R
\name{load_all_books}
\alias{load_all_books}
\title{load_all_books}
\usage{
load_all_books()
}
\value{
Dataframe
}
\description{
Retorna um dataframe com todos os livros contidos no pacote literaturaBR.
}
\examples{
# load_all_books()
}
