% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_doc.R
\docType{data}
\name{memorias_de_um_sargento_de_milicias}
\alias{memorias_de_um_sargento_de_milicias}
\title{Memorias de um Sargento de Milicias - Manuel Antonio de Almeira}
\format{Um dataframe com 1451 linhas e 5 variaveis:
\describe{
  \item{\code{book_name}}{character. Nome do livro.}
  \item{\code{chapter_name}}{character. Capitulo.}
  \item{\code{url}}{character. Link para capitulo no Wikisource.}
  \item{\code{paragraph_number}}{integer. Numero do paragrafo.}
  \item{\code{text}}{character. Texto do paragrafo.}
}}
\usage{
memorias_de_um_sargento_de_milicias
}
\description{
Um dataset que contem o texto integral do livro Memorias de um Sargento de Milicias,
de Manuel Antonio de Almeira,
publicado em 1853.
}
\keyword{datasets}
