% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_doc.R
\docType{data}
\name{memorias_postumas_bras_cubas}
\alias{memorias_postumas_bras_cubas}
\title{Memorias Postumas de Bras Cubas - Machado de Assis}
\format{Um dataframe com 1036 linhas e 5 variaveis:
\describe{
  \item{\code{book_name}}{character. Nome do livro.}
  \item{\code{chapter_name}}{character. Capitulo.}
  \item{\code{url}}{character. Link para capitulo no Wikisource.}
  \item{\code{paragraph_number}}{integer. Numero do paragrafo.}
  \item{\code{text}}{character. Texto do paragrafo.}
}}
\usage{
memorias_postumas_bras_cubas
}
\description{
Um dataset que contem o texto integral do livro Memorias Postumas de Bras Cubas, de Machado de Assis,
publicado em 1881.
}
\keyword{datasets}
