% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books_doc.R
\docType{data}
\name{noite_na_taverna}
\alias{noite_na_taverna}
\title{Noite na Taverna - Álvares de Azevedo}
\format{A data frame with 612 rows and 5 variables:
\describe{
  \item{\code{book_name}}{character. Nome do livro.}
  \item{\code{chapter_name}}{character. Capitulo.}
  \item{\code{url}}{character. Link para capitulo no Wikisource.}
  \item{\code{paragraph_number}}{integer. Numero do paragrafo.}
  \item{\code{text}}{character. Texto do paragrafo.}
}}
\source{
\url{https://pt.wikisource.org/wiki/Noite_na_Taverna}
}
\usage{
noite_na_taverna
}
\description{
Título: Noite na Taverna
Autor: Álvares de Azevendo
Ano: 1855
}
\keyword{datasets}
